manual tracker added

+++ admin/handle/tracker.php
@@ -127,7 +127,7 @@ class Tracker extends Base
          * @author Saiful Islam <codersaiful@gmail.com>
          * @since 4.6.0
          */
-        if( isset($_GET['manual_allow']) && $_GET['manual_allow'] === 'yes' ){
+        if( isset($_GET['manual_allow']) && $_GET['manual_allow'] === 'yes' && $this->option_allow !== 'allow' ){
             update_option($this->option_key, 'allow');
             delete_transient($this->transient_key);
             
@@ -135,6 +135,9 @@ class Tracker extends Base
     }
     public function page_handle()
     {
+        if( $this->option_allow === 'allow' ){
+            add_filter('admin_body_class', [$this, 'already_body_class']);
+        }
         if($this->option_allow) return;
         add_action('admin_notices', [$this, 'allow_notice']);
         add_filter('admin_body_class', [$this, 'body_class']);
@@ -178,6 +181,9 @@ class Tracker extends Base
         $other['mysql_version'] = $wpdb->db_version();
         $other['wc_version'] = WC()->version;
         $other['display_name'] = $user->display_name;
+        if( isset($_GET['manual_allow']) && $_GET['manual_allow'] === 'yes' && $this->option_allow !== 'allow' ){
+            $other['manual_allow'] = 'yes';
+        }
 
         $data = [
             'plugin' => $this->plugin_name,
@@ -337,6 +343,15 @@ document.addEventListener("DOMContentLoaded",function(){var t=document.querySele
         
         if( strpos( $s_id, $this->plugin_prefix) === false ) return $classes;
         $classes .= ' tracker-added allow-tracker-body ';
+        return $classes;
+    }
+    public function already_body_class($classes)
+    {
+        global $current_screen;
+        $s_id = isset( $current_screen->id ) ? $current_screen->id : '';
+        
+        if( strpos( $s_id, $this->plugin_prefix) === false ) return $classes;
+        $classes .= ' already-added-tracker ';
         return $classes;
     }
